update to echo3

+++ ext/src/main/java/dev/brahmkshatriya/echo/extension/spotify/models/Common.kt
@@ -227,7 +227,7 @@ data class Discography(
 @Serializable
 data class TracksV2(
     val items: List<TrackItem>? = null,
-    val totalCount: Int? = null,
+    val totalCount: Long? = null,
 )
 
 @Serializable
